Commit of the first version of tracked_ptr.
It satisfies the current test cases.
However, it is not thread-safe yet, so it can not be used
in the RoarVM code just yet. Also, it is not clear yet, whether the
design is sound with respect to passing on those tracked pointers. I
guess the copy semantics need to be implemented explicitly. (TODO)