Changed global process list to interpreter-local process list (1st stage)
We changed the global process list, which is being accessed by all interpreters, to a local list per interpreter. We convert the list on startup time and on shutdown, allowing us to be compatible with older images.
At the moment it does not yet work for multiple interpreters, as their is no workstealing going on. This will be done in the next stage.

Signed-off-by: Reinout Stevens <resteven@vub.ac.be>