Resume processes on a legal scheduler
We verify that a process is resumed on a legal scheduler, instead of scheduling it on the current squeak interpreter and hope that it is stolen one day (a hint: it won't). Processes with a mask of 0 (the idle process?) are scheduled on the current squeak interpreter, even though they are not allowed to run on any scheduler. Not doing this results in wrong / weird behaviour.

Signed-off-by: Reinout Stevens <resteven@vub.ac.be>