Removed active process from the quiescentProcessList
Removed the active process from the process list, which only has to hold the suspended / non-running processes. This is now equivalent to what squeak does.

Also rewrote /added some assertions, and functionality to print a different process than the running process for debugging purposes.

Note that the 'old' images are b0rked, as there are running processes in the process list (even though this should not be the case in our images). The fix is to recreate a renaissance image from the mvc-clean-image.

Signed-off-by: Reinout Stevens <resteven@vub.ac.be>