Fixed workstealing bug + fixed tilera least_significant_one
The tilera patch is still untested, but should work by using the built-in trailing-zeros instruction and converting that value to a value that provides the same semantics as the gcc-one.

The workstealing bug was that the check for a random interpreter that is not equal to the current one had a != instead of a == . Maybe we now get a speedup.

Signed-off-by: Reinout Stevens <resteven@vub.ac.be>